requirements.txt
requests
pymysql
rsa
pyyaml
allure-pytest
flask
redis
jsonpath
openpyxl
pytest
pytest-xdist
pytest-base-url
pytest-rerunfailures
pytest-ordering
pytest-html
